Major changes to 250607 including eqnset and modcouple functionality numerous minor edits also

+++ EEGcode/trunk/dendriticrlist.h
@@ -27,18 +27,18 @@ public:
   void init(Istrm& inputf, PropagNet * ppropagnet, ConnectMat * pconnectmat);  // initialize dendritic response list including initializing localP a pointer table to the correct part of ppropagnet->P data
   void dump(ofstream& dumpf);
   void restart(Istrm& restartf, PropagNet * ppropagnet, ConnectMat * pconnectmat);  // restart dendritic response list including initializing localP a pointer table to the correct part of ppropagnet->P data
-  void stepVa(float timestep); // Integrates each dendritic subpopulation one timestep forward
-  void SumAfferent(float *V); //Sum sub potentials Va matrix to yield potential V for neural population
+  void stepVa(double timestep); // Integrates each dendritic subpopulation one timestep forward
+  void SumAfferent(double *V); //Sum sub potentials Va matrix to yield potential V for neural population
 
 private: 
   DendriticR ** drarray; // Array of pointers to dendritic responses
   int numaff; //number of afferent neural populations connected to this population
   const long numnodes;
   const int popindex; // population index number of the population in which this dendritic response list is contained
-  float ** Va; // pointer to array of pointers to subpotentials from afferent tree
-  float ** dVdt; // pointer to array of pointers to derivatives of subpotential from afferent tree
+  double ** Va; // pointer to array of pointers to subpotentials from afferent tree
+  double ** dVdt; // pointer to array of pointers to derivatives of subpotential from afferent tree
   DendriticR * getdendr(int i); // method to obtain pointer to "i"th dendritic response
-  float **localP; // A local pointer array pointing to relevant part of P data in propagnet
+  double **localP; // A local pointer array pointing to relevant part of P data in propagnet
 };
 
 #endif
