+++ EEGcode/trunk/proplist.h
@@ -18,17 +18,18 @@ using std::endl;
 #include"propag.h"
 #include"waveeqn.h"
 #include"pmap.h"
+#include"eqnset.h"
 #include"qhistorylist.h"
 #include"connectmat.h"
 
 class Proplist {
 public: 
-  Proplist(Istrm& inputf, ofstream& dumpf, int numconnects, long gridsize, float deltat);
+  Proplist(Istrm& inputf, ofstream& dumpf, int numconnects, long gridsize, double deltat);
   ~Proplist();
   void init(Istrm& inputf);
   void dump(ofstream& dumpf);
   void restart(Istrm& restartf);
-  void step(float **Eta, Qhistorylist * pqhistorylist, ConnectMat * pconnectmat); // Transform Q to Eta via stepping forward multiple wave equations
+  void step(double **Eta, Qhistorylist * pqhistorylist, ConnectMat * pconnectmat); // Transform Q to Eta via stepping forward multiple wave equations
 
 private:
   Propag ** propagarray; // Array of pointers to propagator objects
